# lib/api-zod/src/generated/types/courseModule.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artemis Education API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from "./lesson";

export interface CourseModule {
  id: number;
  titleBn: string;
  titleEn: string;
  order: number;
  lessons?: Lesson[];
}