# lib/api-zod/src/generated/types/createPaymentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artemis Education API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePaymentBodyMethod } from "./createPaymentBodyMethod";

export interface CreatePaymentBody {
  courseId: number;
  amount: number;
  method: CreatePaymentBodyMethod;
  transactionId?: string;
  couponCode?: string;
}