# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artemis Education API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from "./activityItem";

export interface DashboardSummary {
  enrolledCourses: number;
  completedCourses: number;
  totalLessonsCompleted: number;
  totalQuizzesTaken: number;
  averageQuizScore?: number;
  currentStreak: number;
  totalPoints: number;
  upcomingLiveClasses?: number;
  recentActivity?: ActivityItem[];
}