# lib/api-zod/src/generated/types/liveClass.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artemis Education API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LiveClass {
  id: number;
  titleBn: string;
  titleEn: string;
  instructor: string;
  instructorAvatar?: string;
  subject: string;
  scheduledAt: Date;
  durationMinutes: number;
  joinUrl?: string;
  isLive: boolean;
  enrolledCount: number;
  courseId?: number;
}