# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artemis Education API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Quiz {
  id: number;
  titleBn: string;
  titleEn: string;
  courseId?: number;
  totalQuestions: number;
  timeLimitMinutes: number;
  passingScore: number;
}